Summary: Incorrect type for Coefficient Code Generation Reviewed by: Chawla, Pankaj; Prokofiev, Pavel G Rurl: https://ir-codecollab.intel.com/ui#review:id=33896 Old/split llvm commit ID: 29a6865b9a508296e77300f6518e3e017253b4d1

+++ llvm/lib/Transforms/Intel_LoopTransforms/HIRCodeGen.cpp
@@ -188,7 +188,7 @@ private:
     Value *IVPairCG(CanonExpr *CE, CanonExpr::iv_iterator IVIt, Type *Ty);
 
     // \brief retutn value for coeff*V
-    Value *CoefCG(int Coeff, Value *V);
+    Value *CoefCG(int64_t Coeff, Value *V);
 
     //\brief returns value for blobCoeff*blob in <blobidx,coeff> pair
     Value *BlobPairCG(CanonExpr *CE, CanonExpr::blob_iterator BlobIt,
@@ -1111,7 +1111,7 @@ Value *HIRCodeGen::CGVisitor::IVPairCG(CanonExpr *CE,
     return CoefCG(CE->getIVConstCoeff(IVIt), IV);
   }
 }
-Value *HIRCodeGen::CGVisitor::CoefCG(int Coeff, Value *V) {
+Value *HIRCodeGen::CGVisitor::CoefCG(int64_t Coeff, Value *V) {
 
   // do not emit 1*iv, just emit IV
   if (Coeff == 1)
